Improve type hints
With the type hints list and dict there were quite a few errors due to the mutability of lists and dictionaries. The types Sequence and Mapping are the read-only equivalents and they are preferred for the type hinting of the function arguments. For the return values it is perfectly fine to use list and dict since those are actually created in the functions.