Follow strict typing rules
A few lines are ignored for now since I couldn't find a solution that conforms to strict typing. The left and right title of an axes can only be accessed with the underscore attributes. The method axes.get_title() will only return the text.
The easiest way to fix the type errors with the transforms used for the text position is probably to directly work with the matrices and use numpy to apply the transformations. But I'm not yet sure if I want to do that...