Set grid body to None if there are no cells
This should not ever happen since a grid will only be created if there is at least one cell. Passing an empty string as the body argument will however result in compilation error. Passing None to completely skip the body is therefore the cleaner solution to make sure that the generated document always compiles.