Refactor drawing of ticks and labels
This is done in preparation for the grid lines. The old/previous version of the function draw-axis-ticks() only handled one specific alignment of the ticks. This worked for ticks and labels that can have either of the outer alignments. For grid lines this would not work since they can only exist or not.