Finally add support for grid lines
I opted for a separate draw function in Typst after all. Drawing a grid is super simple compared to the ticks and labels, and the grid should only be drawn in the interval (0%, 100%). The labels and ticks are also drawn at 0% and 100%.
Matplotlib itself actually showed some weird behavior here. The grid line at 0% was drawn whereas the grid line at 100% was not drawn? Is there some weird clipping happening at 100%?