Implement abc axes labels
The function is only available for manual placement (for now). In matplotlib, the abc labels are done with manual annotations. This is rather difficult to identify as abc labels instead of other annotations...
In proplot/ultraplot this might be possible by catching the abc style when the plot is created for the first time?